Create Week 3 Part 1 Version 2
So for this version, I changed the namespace to std.
Adjusted the spacing of the { } so they always fall on the same column, not a functional change, but something the instructor mentioned looked nice.
And finally, I changed the parameter of the monthly appointment to search for matching Day and Year. It had been set to the Month. Which would have made it only show up once a year on that month. So now "pay your rent" shows up every month on the first.